Improved Fix for Cursor Movement. Breaks Palm Detection.
With this change on top of the previous one, cursor movement while scrolling is totally fixed, as is all two finger movement when crostrackpadscroll is not running. This fix, unfortunately, breaks palm detection and is unviable unless that can be fixed. The behavior with this fix is, however, very similar to how my rMBP trackpad behaves (movement is only allowed with one finger on the trackpad).